refactor RGBA

+++ src/color.cpp
@@ -1,4 +1,4 @@
 #include "stdafx.h"
 #include "color.h"
 
-RGBA::RGBA(int r, int g, int b, int a) : data{r, g, b, a} {}
+RGBA::RGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a) : data{r, g, b, a} {}
